updating build.rs script
no longer relies on make using CC.

It is slow when it has to build the c objects so currently it only builds those
if one of the .c or .h files are updated. NOTE this important to remember since
it will affect when the auto generated files are made.