Moving Pest parser code into another crate.
Cargo test has been taking a annoyingly long time to run.
It looks like something to do with how pest is written causes
the generated code to be recompiled every time the doctests are run.

To fix this I moved the pest logic into its own crate and used cfg
to disable code generation for the doctests.