Updating lock tab to fill up the rest the pagefile
The resent rsm release I pulled in form upstream updated the locktable
size so that if expanded to fill up the page file. (the shared memory
segment is allocated in units of pagefiles so we were already
allocating that memory just not using it.)

NOTE It looks like one of the tests if flaky.
This seems to be a race condition between a few of my unit tests.