Fixing C name conflicts
I determined that the wrong copy of ST_INIT was being called in my tests
due to name coalitions with the old/new copies.

I am surprised that I was not getting a link error regarding having
duplicate symbols.

As a quick/simple fix for this I just change the prefix of for all the
new functions ST(symbol_table)-> TMP (temporary, they well be converted
to rust at some point)