Modifying start test to clean up the mem_segment
I had this before but it was messed up when I introduced the safe start
interface. To make sure the mem_segement is cleaned up we need the
mem_guard's detractor to run, but that only happens if we actually
initialize the mem_segment. This will work for now, but at some point I
should look into requiring the safe api a bit, it might be nice if start
returned the shared meme segment.